Fix bug caused by type fix
