+++ packages/ai/package.json
@@ -35,7 +35,7 @@
     "dev": "rollup -c -w",
     "update-responses": "../../scripts/update_vertexai_responses.sh",
     "testsetup": "yarn update-responses && yarn ts-node ./test-utils/convert-mocks.ts",
-    "test": "run-p --npm-path npm lint test:browser",
+    "test": "run-p --npm-path npm lint type-check test:browser",
     "test:ci": "yarn testsetup && node ../../scripts/run_tests_in_ci.js -s test",
     "test:skip-clone": "karma start",
     "test:browser": "yarn testsetup && karma start",
@@ -44,6 +44,7 @@
     "test:integration:node": "TS_NODE_COMPILER_OPTIONS='{\"module\":\"commonjs\"}' mocha integration/**/*.test.ts --config ../../config/mocharc.node.js",
     "api-report": "api-extractor run --local --verbose",
     "typings:public": "node ../../scripts/build/use_typings.js ./dist/ai-public.d.ts",
+    "type-check": "yarn tsc --noEmit",
     "trusted-type-check": "tsec -p tsconfig.json --noEmit"
   },
   "peerDependencies": {
