+++ packages/ai/src/index.ts
@@ -22,52 +22,16 @@
  */
 
 import { registerVersion, _registerComponent } from '@firebase/app';
-import { AIService } from './service';
 import { AI_TYPE } from './constants';
-import {
-  Component,
-  ComponentContainer,
-  ComponentType,
-  InstanceFactoryOptions
-} from '@firebase/component';
+import { Component, ComponentType } from '@firebase/component';
 import { name, version } from '../package.json';
-import { decodeInstanceIdentifier } from './helpers';
-import { AIError } from './api';
-import { AIErrorCode } from './types';
-import { chromeAdapterFactory } from './methods/chrome-adapter';
 import { LanguageModel } from './types/language-model';
+import { factory } from './factory-browser';
 
 declare global {
   interface Window {
     LanguageModel: LanguageModel;
   }
-}
-
-function factory(
-  container: ComponentContainer,
-  { instanceIdentifier }: InstanceFactoryOptions
-): AIService {
-  if (!instanceIdentifier) {
-    throw new AIError(
-      AIErrorCode.ERROR,
-      'AIService instance identifier is undefined.'
-    );
-  }
-
-  const backend = decodeInstanceIdentifier(instanceIdentifier);
-
-  // getImmediate for FirebaseApp will always succeed
-  const app = container.getProvider('app').getImmediate();
-  const auth = container.getProvider('auth-internal');
-  const appCheckProvider = container.getProvider('app-check-internal');
-
-  return new AIService(
-    app,
-    backend,
-    auth,
-    appCheckProvider,
-    chromeAdapterFactory
-  );
 }
 
 function registerAI(): void {
